+++ cmn-data-svc-client/src/main/java/org/acumos/cds/domain/MLPPeerSubscription.java
@@ -52,6 +52,10 @@ public class MLPPeerSubscription extends MLPTimestampedEntity implements Seriali
 	@Size(max = 36)
 	private String peerId;
 
+	@Column(name = "OWNER_ID", nullable = false, columnDefinition = "CHAR(36)")
+	@Size(max = 36)
+	private String ownerId;
+
 	// JSON
 	@Column(name = "SELECTOR", columnDefinition = "VARCHAR(1024)")
 	@Size(max = 1024)
@@ -81,11 +85,14 @@ public class MLPPeerSubscription extends MLPTimestampedEntity implements Seriali
 	 * 
 	 * @param peerId
 	 *            Peer ID
+	 * @param ownerId
+	 *            User ID, the operator
 	 */
-	public MLPPeerSubscription(String peerId) {
-		if (peerId == null)
+	public MLPPeerSubscription(String peerId, String ownerId) {
+		if (peerId == null || ownerId == null)
 			throw new IllegalArgumentException("Null not permitted");
 		this.peerId = peerId;
+		this.ownerId = ownerId;
 	}
 
 	public Long getSubId() {
@@ -102,6 +109,14 @@ public class MLPPeerSubscription extends MLPTimestampedEntity implements Seriali
 
 	public void setPeerId(String peerId) {
 		this.peerId = peerId;
+	}
+
+	public String getOwnerId() {
+		return ownerId;
+	}
+
+	public void setOwnerId(String userId) {
+		this.ownerId = userId;
 	}
 
 	public String getSelector() {